+++ PainterSkija/src/main/java/net/buildabrowser/babbrowser/painter/skija/SkijaCommandList.java
@@ -30,7 +30,7 @@ public class SkijaCommandList implements SkijaPaintBitMap {
   }
 
   @Override
-  public void draw(Canvas canvas, Paint paint, int x, int y) {
+  public void draw(Canvas canvas, Paint paint, float x, float y) {
     canvas.save();
     canvas.translate(x, y);
     canvas.drawPicture(picture);
